services/stream-processing/requirements.txt
kafka-python
pyspark
numpy